refactor(STADTPULS-385): projects -> sensors on home and in menu
