+++ pages/sensors.tsx
